+++ packages/react/src/layout/Pickers.js
@@ -1,9 +1,8 @@
 import _ from 'lodash/fp'
-import * as F from 'futil-js'
+import F from 'futil-js'
 import React from 'react'
 import { observer } from 'mobx-react'
 import { withStateLens } from '../utils/mobx-react-utils'
-import { value } from '../utils/actout'
 import TextHighlight from './TextHighlight'
 
 export let FilteredPicker = withStateLens({ filter: '' })(
@@ -17,7 +16,7 @@ export let FilteredPicker = withStateLens({ filter: '' })(
       Highlight = TextHighlight,
     }) => (
       <div>
-        <Input {...value(filter)} />
+        <Input {...F.domLens.value(filter)} />
         {_.map(
           ({ value, label }) => (
             <Item key={value} onClick={() => onChange(value)}>